Backwards-compatible django constraints
